requirements.txt
Flask==3.0.0
qrcode[pil]==7.4.2
Pillow==10.1.0